Improves supply credit yield and value of Ore Scanner data disks. Removes broken Warden and Sec Chief armor vests. Fixes Kitchen Gibber (mostly, grabs still error, but drags work). Allowed Kitchen Spike to work on mouse drag and for simple animals (dogs and carp) as well. Improves Plasma Cutter mining ability - no affect to combat. Allows construction of basic crates from alumin(i)um.
Greatly expands emag use mechanics:
Is no longer an easy-swipe. Requires 6 seconds of stationary action to work.
If the action is interrupted, there is a 10% chance that the emag is damaged.
If it is damaged three times, the emag will break regardless of the prior successful uses.
Increases maximum uses to 12 from 10.